requirements.txt
aiohttp
nextcord
pillow==9.3.0
pyperclip